+++ src/modules/growth/infrastructure/container.ts
@@ -1,9 +1,11 @@
+import { metaService } from "@/modules/meta";
 import {
   PrismaUgcRepository,
   PrismaAmbassadorRepository,
   PrismaReferralOrderRepository,
   PrismaDmCampaignRepository,
   PrismaBackInStockRepository,
+  PrismaCommentUnlockRepository,
 } from "./repositories";
 import { makeGrowthService } from "../application/service";
 import { makeGrowthQueries } from "../application/queries";
@@ -13,6 +15,7 @@ const ambassadors = new PrismaAmbassadorRepository();
 const referrals = new PrismaReferralOrderRepository();
 const campaigns = new PrismaDmCampaignRepository();
 const backInStock = new PrismaBackInStockRepository();
+const commentUnlocks = new PrismaCommentUnlockRepository();
 
 export const growthService = makeGrowthService({
   ugc,
@@ -20,6 +23,8 @@ export const growthService = makeGrowthService({
   referrals,
   campaigns,
   backInStock,
+  commentUnlocks,
+  meta: metaService,
 });
 
 export const growthQueries = makeGrowthQueries({
@@ -28,4 +33,5 @@ export const growthQueries = makeGrowthQueries({
   referrals,
   campaigns,
   backInStock,
+  commentUnlocks,
 });